Initial Push to repository
Initial push to repository
Resolved previous issue with hal_flash configuration
Currently the debug UART port is configured for PC06(RX) and PC08 (TX)
These are available on the WSTK as solder through hole points P33 and P31 respectively.

PA9 is currently being used for the NRST pin on the LR11xx module from the expansion header.
PA8 is currently being used for the DIOX pin from the expansion header.