Low Power fix
High power consumption was observed when the BG24 entered EM2.  Was approximately 1mA.
primary source of power draw was the flash chipset that is placed on the WSTK.
The pin that was configured for the scan LED was tied to the Chip Select pin of the flash.  With the LED state set it caused the flash to go into a high power state